+++ .circleci/config.yml
@@ -15,9 +15,9 @@ jobs:
       # based on this: https://discuss.circleci.com/t/circle-2-0-caching-is-too-limited-to-be-very-useful/11694/10
       - restore_cache:
           keys:
-            - spring-boot-messaging-work-queue-booster-{{ .Branch }}-{{ checksum "pom.xml" }}
-            - spring-boot-messaging-work-queue-booster-{{ .Branch }}
-            - spring-boot-messaging-work-queue-booster
+            - messaging-work-queue-example-{{ .Branch }}-{{ checksum "pom.xml" }}
+            - messaging-work-queue-example-{{ .Branch }}
+            - messaging-work-queue-example
 
       - run:
           name: Install Maven dependencies locally
@@ -45,7 +45,7 @@ jobs:
       - save_cache:
           paths:
             - ~/.m2
-          key: spring-boot-messaging-work-queue-booster-{{ .Branch }}-{{ checksum "pom.xml" }}
+          key: messaging-work-queue-example-{{ .Branch }}-{{ checksum "pom.xml" }}
 
       - run:
           name: run tests against s2i build
