--- frontend/src/main/java/io/openshift/booster/messaging/FrontendHeaders.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2018 Red Hat, Inc, and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.openshift.booster.messaging;
-
-public final class FrontendHeaders {
-
-    public static String WORKER_ID = "workerId";
-
-    public static String TIMESTAMP = "timestamp";
-
-    public static String REQUESTS_PROCESSED = "requestsProcessed";
-
-    public static String PROCESSING_ERRORS = "processingErrors";
-
-    public static String UPPERCASE = "uppercase";
-
-    public static String REVERSE = "reverse";
-
-}
