--- frontend/src/main/java/io/openshift/booster/messaging/Request.java
@@ -1,58 +1,0 @@
-/*
- * Copyright 2018 Red Hat, Inc, and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.openshift.booster.messaging;
-
-public class Request {
-
-    private String text;
-
-    private boolean uppercase;
-
-    private boolean reverse;
-
-    public Request() {
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public boolean isUppercase() {
-        return uppercase;
-    }
-
-    public void setUppercase(boolean uppercase) {
-        this.uppercase = uppercase;
-    }
-
-    public boolean isReverse() {
-        return reverse;
-    }
-
-    public void setReverse(boolean reverse) {
-        this.reverse = reverse;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("Request{text=%s, uppercase=%s, reverse=%s}", text, uppercase, reverse);
-    }
-}
