+++ tests/pom.xml
@@ -4,14 +4,14 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>io.openshift.booster</groupId>
-    <artifactId>spring-boot-messaging-work-queue-parent</artifactId>
+    <groupId>dev.snowdrop.example</groupId>
+    <artifactId>messaging-work-queue</artifactId>
     <version>1.5.19-3-SNAPSHOT</version>
   </parent>
 
-  <artifactId>spring-boot-messaging-booster-tests</artifactId>
+  <artifactId>spring-boot-messaging-example-tests</artifactId>
 
-  <name>Spring Boot - Messaging - Work Queue Booster - Integration Tests</name>
+  <name>Spring Boot - Messaging - Work Queue Example - Integration Tests</name>
 
   <dependencies>
     <dependency>
@@ -93,12 +93,12 @@
             </configuration>
             <dependencies>
               <dependency>
-                <groupId>io.openshift.booster</groupId>
+                <groupId>dev.snowdrop.example</groupId>
                 <artifactId>spring-boot-messaging-work-queue-frontend</artifactId>
                 <version>${project.version}</version>
               </dependency>
               <dependency>
-                <groupId>io.openshift.booster</groupId>
+                <groupId>dev.snowdrop.example</groupId>
                 <artifactId>spring-boot-messaging-work-queue-worker</artifactId>
                 <version>${project.version}</version>
               </dependency>
