--- worker/src/main/java/io/openshift/booster/messaging/WorkerApplication.java
@@ -1,48 +1,0 @@
-/*
- * Copyright 2018 Red Hat, Inc, and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.openshift.booster.messaging;
-
-import javax.jms.ConnectionFactory;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jms.DefaultJmsListenerContainerFactoryConfigurer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.jms.annotation.EnableJms;
-import org.springframework.jms.config.DefaultJmsListenerContainerFactory;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication
-@EnableJms
-@EnableScheduling
-public class WorkerApplication {
-
-    public static void main(String... args) {
-        SpringApplication.run(WorkerApplication.class, args);
-    }
-
-    @Bean
-    public DefaultJmsListenerContainerFactory jmsListenerContainerFactory(ConnectionFactory connectionFactory,
-            DefaultJmsListenerContainerFactoryConfigurer configurer) {
-        DefaultJmsListenerContainerFactory listenerFactory = new DefaultJmsListenerContainerFactory();
-        configurer.configure(listenerFactory, connectionFactory);
-        listenerFactory.setTransactionManager(null);
-        listenerFactory.setSessionTransacted(false);
-        return listenerFactory;
-    }
-
-}
